// WebUI/src/services/client/purchaseService.ts
export interface PurchaseEntity {
    CustomerID?: string | undefined;
    Id?: number | undefined;
    PurchaseDate?: Date | undefined;
}

export interface PurchaseRequest {
    purchase?: PurchaseEntity | undefined;
    stickers?: StickerEntity[] | undefined;
}

export interface StickerEntity {
    Description?: string | undefined;
    Id?: number | undefined;
    Price?: number | undefined;
    URL?: string | undefined;
    description?: string | undefined;
    id?: number | undefined;
    price?: number | undefined;
    url?: string | undefined;
}